moving bevy related structs to a different bundle
