+++ workspaces/maploading/src/util/map/mod.rs
@@ -1,8 +1,9 @@
 pub mod loader;
 mod data;
 mod util;
+mod bundles;
 mod entity_builder;
 pub use data::MapFile;
-pub use entity_builder::Collider;
+pub use bundles::Collider;
 
 pub const MAP_NAME: &str = "level1";
